pyproject.toml
[tool.poetry]
name = "odoo-openupgrade-wizard"
version = "1.0.3"
description = "CLI tool to manage Odoo Major Upgrades"
authors = [
  "Sylvain LE GAL <sylvain.legal@grap.coop>",
  "Rémy TAYMANS <remy@coopiteasy.be>",
  "Simon MAILLARD <simon@ogesta.fr>",
]
maintainers = [
  "Sylvain LE GAL <sylvain.legal@grap.coop>",
  "Rémy TAYMANS <remy@coopiteasy.be>",
]
readme = "README.md"
repository = "https://gitlab.com/odoo-openupgrade-wizard/odoo-openupgrade-wizard"
keywords = ["cli", "odoo", "openupgrade"]
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Intended Audience :: Developers",
    "Development Status :: 4 - Beta",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Odoo",
]

[project.urls]
Repository = "https://gitlab.com/odoo-openupgrade-wizard/odoo-openupgrade-wizard"

[tool.poetry.scripts]
oow = "odoo_openupgrade_wizard.cli.cli:main"
odoo-openupgrade-wizard = "odoo_openupgrade_wizard.cli.cli:main"

[tool.poetry.dependencies]
python = ">=3.9,<4.0.0"
click = "^7.0"
click-loglevel = "^0.4"
docker = "^7.0"
importlib-resources = "^5.4"
git-aggregator = "^2.1"
GitPython = "^3.1"
jinja2 = "^3.0"
loguru = "^0.6"
odoorpc = "^0.10.1"
plumbum = "^1.7"
pygount = "^1.4"
pyyaml = "^6.0.0"
single-source = "^0.3"

[tool.poetry.group.dev.dependencies]
pytest = [
    {version = "<=6.1.2", python = "<3.10"},
    {version = ">=6.2.5", python = ">=3.10"}
]
pytest-mock = "3.6.1"
pytest-cov = "*"
safety = "*"
pylint = "*"
tox = "*"
towncrier = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.towncrier]
name = "odoo-openupgrade-wizard"
package = "odoo_openupgrade_wizard"
package_dir = "."
filename = "CHANGES.rst"
directory = "newsfragments"
wrap = true

[[tool.towncrier.type]]
directory = "feature"
name = "Features"
showcontent = true

[[tool.towncrier.type]]
directory = "bugfix"
name = "Bugfixes"
showcontent = true

[[tool.towncrier.type]]
directory = "removal"
name = "Removals"
showcontent = true

[[tool.towncrier.type]]
directory = "doc"
name = "Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "misc"
name = "Misc"
showcontent = true
